backend/requirements.txt
Jinja2==3.1.6
xhtml2pdf==0.2.15
python-dateutil==2.8.2
reportlab==4.0.4
